requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Machine Learning for anomaly detection
scikit-learn>=1.1.0

# AI/LLM integration
openai>=1.0.0

# Time series analysis (for future use)
statsmodels>=0.13.0

# Environment configuration
python-dotenv>=0.19.0

# Optional: For better data visualization (if needed)
matplotlib>=3.5.0
seaborn>=0.11.0